chore: move action package into the process package
